chore(motor): change id type from uint16_t to uint32_t

+++ base/motor/RM3508.cpp
@@ -9,7 +9,7 @@ namespace esf
 {
 namespace base
 {
-RM3508::RM3508(uint16_t id, float reduration_ratio, Direction direction)
+RM3508::RM3508(uint32_t id, float reduration_ratio, Direction direction)
     : Motor<RM3508, RM3508State, uint16_t>(id, reduration_ratio, direction)
     , m_current_param(0.3f / (3591.0f / 17.0f) * reduration_ratio)
 {
@@ -83,12 +83,12 @@ template <> EsfStatus RM3508Helper<FDCAN>::_decodeMessageImpl(FDCAN::Message &me
     auto last_rotor_position = motor.state().rotor_position; // 记录上次转子位置
     // 记录原始信息，RM3508 原始信息均为转子信息
     motor.state().rotor_position = 1.0f * static_cast<float>(motor.m_direction) *
-                                    static_cast<float>((static_cast<uint16_t>(data[0]) << 8) | data[1]) * 2.0f * M_PI /
-                                    8191.0f;
+                                   static_cast<float>((static_cast<uint16_t>(data[0]) << 8) | data[1]) * 2.0f * M_PI /
+                                   8191.0f;
     motor.state().rotor_velocity = 1.0f * static_cast<float>(motor.m_direction) *
-                                    static_cast<float>((static_cast<int16_t>(data[2]) << 8) | data[3]);
+                                   static_cast<float>((static_cast<int16_t>(data[2]) << 8) | data[3]);
     motor.state().rotor_current = 1.0f * static_cast<float>(motor.m_direction) *
-                                   static_cast<float>(static_cast<int16_t>(data[4] << 8) | data[5]) / 16384.0f * 10.0f;
+                                  static_cast<float>(static_cast<int16_t>(data[4] << 8) | data[5]) / 16384.0f * 10.0f;
     motor.state().temperature = static_cast<float>(data[6]);
     // 计算输出轴数据
     auto rotor_position_diff = motor.state().rotor_position - last_rotor_position;
